# app/static/swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "Cinema API"
  description: "This is a Cinema API where you can execute CRUD operations on 
  movies, actors, genres, sessions, customers and auditoriums"
servers:
  - url: http://127.0.0.1:5000
...
paths:
  /actors:
    get:
      tags:
        - "Actors"
      summary: "Get actors information"
      parameters:
        - name: "firstname"
          in: "query"
          description: "Firstname of actor"
          required: false
          schema:
            type: "string"
        - name: "lastname"
          in: "query"
          description: "Lastname of actor"
          required: false
          schema:
            type: "string"
        - name: "offset"
          in: "query"
          description: "Offset"
          required: false
          schema:
            type: "string"
        - name: "limit"
          in: "query"
          description: "Limit"
          required: false
          schema:
            type: "string"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorsOut'
    post:
      security:
        - bearerAuth: []
      tags:
        - "Actors"
      summary: "Create new actor"
      description: "This can only be done by the logged in admin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActorIn'
      responses:
        '201':
          description: "Successful Operation"
          content:
            application/json:
              example:
                id: "5"
        '400':
          description: "Wrong input data"
          content:
            application/json:
              example:
                message: "Please, specify 'firstname' and 'lastname'"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
  /actors/{id}:
    get:
      tags:
        - "Actors"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of actor"
          required: true
          schema:
            type: "integer"
      summary: "Get actor by id"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorOut'
        '404':
          description: "Actor not found"
          content:
            application/json:
              example:
                message: "Actor not found"
    patch:
      security:
        - bearerAuth: []
      tags:
        - "Actors"
      summary: "Update actor's information"
      description: "This can only be done by the logged in admin"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of actor"
          required: true
          schema:
            type: "integer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActorIn'
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Updated"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong actor's id"
          content:
            application/json:
              example:
                message: "Actor not found"
    delete:
      security:
        - bearerAuth: []
      tags:
        - "Actors"
      summary: "Delete actor"
      description: "This can only be done by the logged in admin"
      parameters:
        - name: "id"
          in: "path"
          description: "Id of the actor"
          required: true
          schema:
            type: "integer"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Deleted"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong actor's id"
          content:
            application/json:
              example:
                message: "Actor not found"
  /auditorium:
    get:
      tags:
        - "Auditoriums"
      summary: "Get auditoriums information"
      parameters:
        - name: "offset"
          in: "query"
          description: "Offset"
          required: false
          schema:
            type: "string"
        - name: "limit"
          in: "query"
          description: "Limit"
          required: false
          schema:
            type: "string"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditoriumsOut'
    post:
      security:
        - bearerAuth: []
      tags:
        - "Auditoriums"
      summary: "Create new auditorium"
      description: "This can only be done by the logged in admin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditoriumIn'
      responses:
        '201':
          description: "Successful Operation"
          content:
            application/json:
              example:
                id: "5"
        '400':
          description: "Wrong input data"
          content:
            application/json:
              example:
                message: "Please, specify 'seats'"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
  /auditorium/{id}:
    get:
      tags:
        - "Auditoriums"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of auditorium"
          required: true
          schema:
            type: "integer"
      summary: "Get auditorium by id"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditoriumOut'
        '404':
          description: "Auditorium not found"
          content:
            application/json:
              example:
                message: "Auditorium not found"
    patch:
      security:
        - bearerAuth: []
      tags:
        - "Auditoriums"
      summary: "Update auditorium's information"
      description: "This can only be done by the logged in admin"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of actor"
          required: true
          schema:
            type: "integer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditoriumIn'
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Updated"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong auditorium's id"
          content:
            application/json:
              example:
                message: "Auditorium not found"
    delete:
      security:
        - bearerAuth: []
      tags:
        - "Auditoriums"
      summary: "Delete auditorium"
      description: "This can only be done by the logged in admin"
      parameters:
        - name: "id"
          in: "path"
          description: "Id of the auditorium"
          required: true
          schema:
            type: "integer"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Deleted"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong auditorium's id"
          content:
            application/json:
              example:
                message: "Auditorium not found"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ActorOut:
      type: "object"
      properties:
        id:
          type: "integer"
          example: 2
        firstname:
          type: "string"
          example: "Emma"
        lastname:
          type: "string"
          example: "Stone"
        movies:
          type: "array"
          example: []
    ActorsOut:
      type: "array"
      items:
        $ref: '#/components/schemas/ActorOut'
    ActorIn:
      type: "object"
      properties:
        firstname:
          type: "string"
          example: "Emma"
        lastname:
          type: "string"
          example: "Stone"
    AuditoriumOut:
      type: "object"
      properties:
        id:
          type: "integer"
          example: 2
        seats:
          type: "integer"
          example: 60
        movie_sessions:
          type: "array"
          example: []
    AuditoriumsOut:
      type: "array"
      items:
        $ref: '#/components/schemas/AuditoriumOut'
    AuditoriumIn:
      type: "object"
      properties:
        seats:
          type: "integer"
          example: 60
...